moved this to the pgq-parser branch on my fork
Property Graph Query (PGQ) is the main novelty in SQL:2023

THIS CODE COMPILES, BUT DOES NOT WORK YET
- there will be bugs in there
- the parsed PGQ is just parsed, not yet transformed, bound and rewritten

This parser is rather complete, but the following features are not supported:
- simplified path syntax
- undirected edges (~) mostly a GQL feature anyway. Because SQL/PGQ does not have a way to even define such edges.
- ALTER PROPERTY GRAPH.
- DEFAULT LABEL in CREATE PROPERTY GRAPH
  as DuckDB requires all element tables to have unique labels.
- the % wildcard while MATCH-ing elements, used for testing if an element has a label (DuckDB requires elements to have a label)

there will be three language extensions:
- COST(var_name_ident) in a COLUMNS() list of GRAPH_TABLE: var_name_ident must be a path variable. This returns path cost.
- ELEMENT_ID(var_name_ident) in a COLUMNS() list of GRAPH_TABLE: var_name_ident must be a path variable. This returns path cost.
- LABEL global_label IN col LABEL label_list in CREATE PROPERTY GRAPH. In addition to global_label, which all elements in the table get, column 'col' contains an integer interpreted as bitmap. The element also has the x-th label of the label_list iff the x-th bit of that integer is set to 1. This allows flexible LABEL memberships from a single table.

COST and ELEMENT_ID will be parsed as a_expr expressions in the column list and we treat them differently in the transformation.

The constraint that a label occurs only in one element table of a property graph will be enforced in DuckPGQ to avoid having to scan multiple tables for a labeltest.

We detected an amibuity in the SQL syntax resulting from adding SQL/PGQ and ran into shift/reduce issues that had to be fixed:

MATCH <graph-pattern> WHERE <cond> COLUMNS( <list> )

but COLUMNS(list) is a SQL expression (a_expr) and SQL supports infix and suffix operators, like via these grammar rules:

 qual_Op b_expr
 b_expr qual_Op

So in order to parse SQL/PGQ WHERE conditions (as well as COST expressiobs) we had to create a copy of a_expr (in pgq.y that holds all the PGQ stuff) renamed pgq_expr, but with these rules from select.y removed:

 2282 | qual_Op b_expr                                        %prec Op
 2284 | b_expr qual_Op                                        %prec POSTFIXOP
 2186 | a_expr subquery_Op sub_type select_with_parens        %prec Op
 2197 | pgq_expr subquery_Op sub_type '(' pgq_expr ')'                %prec Op
 2204 | DEFAULT
 2218 | COLUMNS '(' '*' opt_except_list opt_replace_list ')'
 2227 | COLUMNS '(' Sconst ')'